+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * None.
 
 ### Enhancements
-* None.
+* Added support for `User.logOut()` ([#245](https://github.com/realm/realm-kotlin/issues/245))
 
 ### Fixed
 * Gradle metadata for pure Android projects. Now using `io.realm.kotlin:library-base:<VERSION>` should work correctly.
@@ -25,6 +25,7 @@
 * Updated to Android targetSdk 31.
 * Updated to Android compileSdk 31. 
 * Updated to Android Build Tools 31.0.0.
+* Updated to Realm Core commit: 2183e09f40a3ce31be0f93d0665b785b5c52e975
 
 
 ## 0.7.0 (2021-10-31)
@@ -33,7 +34,7 @@
 * None.
 
 ### Enhancements
-* Basic MongoDB Realm sync support: 
+* Basic MongoDB Realm sync support:
   * Enabled by using library dependency `io.realm.kotlin:library-sync:<VERSION>`
   * Build `AppConfiguration`s through `AppConfiguration.Builder(appId).build()`
   * Linking your app with a MongoDB Realm App through `App.create(appConfiguration)`